requirements.txt
albumentations==2.0.7
matplotlib==3.10.3
numpy==2.2.6
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
pandas==2.2.3
Pillow==11.2.1
Requests==2.32.3
scikit_learn==1.6.1
seaborn==0.13.2
tqdm==4.67.1
